Fixing error LED flashes for SD write failures
The initFileName() function contained old references to pins
for the error LED, used when the program could not create a
new file of the name stored in the 'filename' variable. Now it
flashes the ERRLED (red).

Also added code in the writeToSD() and initFileName() functions
to first turn off the error LED if it is still on from some
previous error.